+++ cypress/integration/admin/professions/add-a-profession.spec.ts
@@ -78,6 +78,31 @@ describe('Adding a new profession', () => {
         cy.get('button').contains(buttonText).click();
       });
 
+      cy.translate('professions.form.headings.qualificationInformation').then(
+        (heading) => {
+          cy.get('body').should('contain', heading);
+        },
+      );
+
+      cy.get('textarea[name="level"]').type('An example Qualification level');
+      cy.get(
+        'input[name="methodToObtainQualification"][value="others"]',
+      ).check();
+      cy.get('textarea[name="otherMethodToObtainQualification"]').type(
+        'Another method',
+      );
+      cy.get(
+        'input[name="mostCommonPathToObtainQualification"][value="generalSecondaryEducation"]',
+      ).check();
+      cy.get('input[name="duration"]').type('4.0 Years');
+      cy.get(
+        'input[name="mandatoryProfessionalExperience"][value="1"]',
+      ).check();
+
+      cy.translate('app.continue').then((buttonText) => {
+        cy.get('button').contains(buttonText).click();
+      });
+
       cy.translate('professions.form.headings.checkAnswers').then((heading) => {
         cy.get('body').should('contain', heading);
       });
@@ -95,6 +120,18 @@ describe('Adding a new profession', () => {
       });
       cy.get('body').should('contain', 'An example activity');
       cy.get('body').should('contain', 'A description of the regulation');
+
+      cy.get('body').should('contain', 'An example Qualification level');
+      cy.get('body').should('contain', 'Another method');
+      cy.translate(
+        'professions.form.radioButtons.methodsToObtainQualification.generalSecondaryEducation',
+      ).then((method) => {
+        cy.get('body').should('contain', method);
+      });
+      cy.get('body').should('contain', '4.0 Years');
+      cy.translate('app.yes').then((yes) => {
+        cy.get('body').should('contain', yes);
+      });
 
       cy.translate('professions.form.button.create').then((buttonText) => {
         cy.get('button').contains(buttonText).click();